few things
added file creation of summary.txt which creates a summary of crawled links and the date in which it was created, made ReadUrl() return url as a string and put it into the crawler parameter. Program now clears folder on start to avoid overwriting files everytime. started on same domain crawl method(not implemented)